# site/api/openapi.yaml
openapi: 3.0.3
info:
  title: MediaNest API
  description: |
    MediaNest API provides comprehensive media management, Plex integration, YouTube downloading, and system administration capabilities.

    ## Authentication
    - **Primary**: Plex OAuth authentication
    - **Secondary**: Password-based authentication for admin users
    - **Token**: JWT tokens stored in HTTP-only cookies
    - **CSRF**: Required for state-changing operations

    ## Features
    - Media search and request management
    - Plex server integration and library browsing
    - YouTube video downloading with progress tracking
    - Real-time WebSocket events
    - Comprehensive admin panel
    - Performance monitoring and optimization

  version: 1.0.0
  contact:
    name: MediaNest API Support
  license:
    name: MIT
servers:
  - url: http://localhost:4000/api/v1
    description: Development server
  - url: https://api.medianest.app/api/v1
    description: Production server
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /plex/search:
    get:
      tags:
        - Plex
      summary: Search Plex content
      description: Search across all Plex libraries
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlexSearchResponse'
  /plex/recently-added:
    get:
      tags:
        - Plex
      summary: Get recently added items
      description: Get recently added items across all libraries
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Recently added items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlexRecentlyAddedResponse'

  # YouTube Downloads
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: token
...
  schemas:
...
    PlexSearchResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            results:
              type: array
              items:
                type: object
    PlexRecentlyAddedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            items:
              type: array
              items:
                type: object

    # YouTube Schemas
...